% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayGMRF.R
\name{BayGMRF}
\alias{BayGMRF}
\title{Bayesian latent GMRF main function}
\usage{
BayGMRF(Y, X, Q.list, mcmc, options, hyper, model = "Gaussian", E = 1,
  Q.X = 1)
}
\arguments{
\item{Y}{X}

\item{X}{X}

\item{mcmc}{mcmc options}

\item{options}{options}

\item{hyper}{hyperparameters}

\item{model}{model}

\item{E}{E}

\item{nn}{list of neighbours}
}
\value{
list of samples
}
\description{
Bayesian latent GMRF main function
}
