% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_mask.R
\name{img.reg.mask}
\alias{img.reg.mask}
\title{Plot 2D image from coordinate data}
\usage{
img.reg.mask(x, coords, dims = apply(coords, 2, max))
}
\arguments{
\item{x}{values}

\item{coords}{coordinates (2D)}

\item{dims}{dimension of image}
}
\description{
Plot 2D image from coordinate data
}
