% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make3Dfromcoords.R
\name{make.3D.from.coords}
\alias{make.3D.from.coords}
\title{Make a 3D array from a list of coodinates}
\usage{
make.3D.from.coords(x, coords, dims)
}
\arguments{
\item{x}{list of values}

\item{coords}{list of coordinates}

\item{dims}{first two dimensions}
}
\description{
Make a 3D array from a list of coodinates
}
