% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQ.R
\name{makeQ}
\alias{makeQ}
\title{Make precision matrix}
\usage{
makeQ(n, type, params = NULL)
}
\arguments{
\item{n}{dimension of Q}

\item{type}{"rw1" or "I" or "rw1+I"}
}
\value{
sparseMatrix (Matrix package)
}
\description{
Make precision matrix
}
