% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Plot a map}
\usage{
map(data, map, cutpoints = seq(min(data), max(data), length = 256),
  autoscale = FALSE, legend = TRUE, append = FALSE)
}
\arguments{
\item{data}{data}

\item{map}{a list of the coordinates of region border}

\item{cutpoints}{cuts for data scale}

\item{autoscale}{scale data to [0,1]}

\item{legend}{plot a legend}

\item{append}{add map to existing plot}
}
\description{
Plot a map
}
\examples{
data(scotland)
map(1:52,scotland.shape)
}
