% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_regular.R
\name{prepare.regular}
\alias{prepare.regular}
\title{Prepare data from regular array with mask}
\usage{
prepare.regular(data, mask)
}
\arguments{
\item{data}{data}

\item{mask}{mask}
}
\value{
list with data as vector, matrix of coordinates, matrix of neighbour pairs
}
\description{
Prepare data from regular array with mask
}
