% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soi.R
\name{soi}
\alias{soi}
\title{Title}
\usage{
soi(Y, X, Z, Q, all = TRUE, goldsmith = FALSE,
  mcmc.controls = list(nr.iterations = 1000, burnin = 100), plot = TRUE)
}
\arguments{
\item{Y}{}

\item{X}{}

\item{Z}{}

\item{Q}{}

\item{all}{}

\item{goldsmith}{}

\item{mcmc.controls}{}

\item{plot}{}
}
\description{
Title
}
