% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.tau}
\alias{update.tau}
\title{Update tau}
\usage{
\method{update}{tau}(i, tau, beta, Q, Q.X, d, hyper, options)
}
\arguments{
\item{i}{}

\item{tau}{}

\item{beta}{}

\item{Q}{}

\item{Q.X}{}

\item{d}{}

\item{hyper}{}

\item{options}{}
}
\description{
Update tau
}
