% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedQ.R
\name{weightedQ}
\alias{weightedQ}
\title{make GMRF precision matrix with weights}
\usage{
weightedQ(weights, nn)
}
\arguments{
\item{weights}{weights}

\item{nn}{matrix with pair of neighbours per row.}
}
\value{
sparseMatrix of dimension max(nn)
}
\description{
make GMRF precision matrix with weights
}
\examples{
Q<-weightedQ(1:4,matrix(c(1,1,2,3,2,3,3,4),ncol=2))
print(Q)
}
